backend/requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic>=2.9.2
alembic
python-dotenv
psycopg[binary]
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
python-multipart
fpdf2
